requirements.txt
pandas
pytest
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
anytree
openai==1.55.3
icecream
google-cloud-bigquery>=3.11.0
pytz
db-dtypes
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.8.0